EclMaterialLawManager: move (most) code to separate compile unit
